feat: add php session

+++ php/profile/registration_delete.php
@@ -13,24 +13,40 @@
 <body>
     <?php
     include "../connect.php";
-    // variables
-    $eventid = $_GET["eventid"];
-    $participantid = $_GET["participantid"];
+    session_start();
+    if (isset($_SESSION["username"])) :
+        $userid = $_SESSION["userid"];
+        $sessionusername = $_SESSION["username"];
+        $role = $_SESSION["role"];
 
-    // sql
-    $sql = "DELETE FROM registrations WHERE participantid = '$participantid' AND eventid = '$eventid'";
-    $result = mysqli_query($con, $sql) or die("Error deleting data due to " . mysqli_error($con));
+        // variables
+        $eventid = $_GET["eventid"];
+        $participantid = $_GET["participantid"];
 
-    // response
-    if ($result) {
+        // sql
+        $sql = "DELETE FROM registrations WHERE participantid = '$participantid' AND eventid = '$eventid'";
+        $result = mysqli_query($con, $sql) or die("Error deleting data due to " . mysqli_error($con));
 
-        echo "You have successfully left the event.";
-        header("location:profile.php#events");
-    } else {
-        echo "Failed to leave the event.";
-        header("location:profile.php#events");
-    }
-    ?>
+        // response
+        if ($result) {
+
+            echo "You have successfully left the event.";
+            header("location:profile.php#events");
+        } else {
+            echo "Failed to leave the event.";
+            header("location:profile.php#events");
+        }
+    else : ?>
+        <div class="container">
+            <main>
+                <div class="py-5 text-center">
+                    <h2>Error</h2>
+                    <p>Unauthorized access.</p>
+                    <a class="btn btn-primary" href="../signin/signin.php">Sign In</a>
+                </div>
+            </main>
+        </div>
+    <?php endif; ?>
 
     <script src="../../js/colormodes.js"></script>
     <script src="../../js/navbar.js"></script>
